SceneApp refeactor out types
